Fix #10 Provide FQCN generator facility
The ``utils.get_fqcn`` function allows to fetch the FQCN of a ``type``.
With the implementation, I realised there was a problem with
registering and encoder using only the FQCN, mostly if we want to
serialise object from different language (say from JS). Strictly
coupling the `awesojsontype` attribute with the python FQCN was a nad
idea. Instead, I'll use the ``type`` itself for registration, and use
it's FQCN as the defaut `awesojsontype` value, while allowing wo change
that with a 3rd parameter.

Also, `awesojsontype` attribute == textual type identifier.